Make collation and char set optional in create db
Added code to make the collation and character set field in the
create new database form hidden if the datastore is not a mysql
compatible datastore.

Removed the help from the form template since the fields are now
dynamic.  There is still help on the individual fields when visible.

Updated unit tests to account for this change.

Change-Id: I9fdc9a3d4c77a2118b46f70595f1bf94ad451a2b
Closes-Bug: DBAAS-2187