Add support for single instance cluster backup
Add a setting to be enabled in the local settings that takes a
list of datastore names that allow single instance backup in a
cluster to be performed.

Add a backup action to the cluster instances panel.

Added url parameter to the create backup to specify whether to
include_clustered instances or not.

Modified the api list to take the include_clustered flag.

Updated unit tests.

Change-Id: I58ddda5c3f6ab31969d20b4a39cd2bbd1e29b373
Closes-Bug: DBAAS-2165