Add support for cluster upgrade
Added a row menu action for cluster upgrade only if there are
multiple datastore versions for the cluster datastore.

The action will bring up a dialog showing the current version
and a dropdown listing all the datastore versions (with the
current version filterd out of the list).

Added a new trove api for cluster_upgrade.

Change-Id: I2cb945b1ed5bb694edf3fd9d7f5c4ef7f1c23171
Closes-Bug: DBAAS-2221