Add support for user roles
Added the roles field to the create new user dialog.  The roles
field is enabled only if the following is added to the
local_settings file.

TROVE_ENABLE_USER_ROLES = ["couchbase"]

Added the roles column to the user list table.

Modified the api to add the role and updated the unit tests.

Change-Id: I523c9c0365f4b5894152a52339c1d174df46c85a
Closes-Bug: DBAAS-2241