Horizon: Check table data before API call
When force deleting an instance or cluster, the UI might still show
the instance in ERROR state. If force delete is selected again the
API will throw a "Resource cannot be found" error.

Add check to see if the instance/cluster is still in the table data
before calling the API.

Change-Id: I05d83f881e1b4074170f3e5253433b3cae69372a
Closes-Bug: DBAAS-2213