Horizon: Cannot apply boolean parameters
Attempting to apply a boolean parameter to a configuration group results
in a "Error applying changes" error. When the parameter is added the value
(true/false) is treated as a string instead of a boolean.

Add check to see if parameter is boolean type and convert value from string
to boolean.

Change-Id: Ibc5b41c820131ed0ed15f31479cb4e4cb7517a30
Closes-Bug: DBAAS-2304