Add cluster support for user, database and config
Add menu options on the cluster list panel for managing users,
managing databases, attaching/detaching configuration groups
and restarting the cluster.

The manage user action will take the user to a users list panel.
On this panel the new users can be created.  User information can
be updated.  Users can be deleted.  Database access for a user
can is done through a manage access panel that displays the
databases the user has access to.  Access can be granted or revoked.

The manage database action will take the user to a databases list
panel.  On this panel the new databases can be created.  Databases
can be deleted from this panel.

The attach configuration menu option will be visible only if the
cluster does not have an associated configuration group.  Selecting
the action will bring up a dialog where the user can pick the
configuration group to attach to the cluster.

The detach configuration menu option will be visible only if the
cluster has a configuration group associated with it.  Selecting
the action will cause the configuration group to be detached from
the cluster.

The cluster details page is augmented with the configuration id
if a configuration group is attached to the cluster.

The configurations instance page now has an added column to
indicate the cluster id that the instance belongs to.  This
cluster id is a link that will take you to the details of the
cluster.

Added new api calls and unit tests.

Change-Id: Ie37802a6d5cde99eeb4423a40dddf6fa20c5df89
Closes-Bug: DBAAS-2301