Horizon: List users for cassandra datastore
On the Horizon Dashboard when listing the users for a casssandra
datastore instance the following error is returned "Unable to get
user access data"

The list user access API call does not return hosts for cassandra,
which is the correct behaviour. Add condition to check if host is
being returned and handle the condition accordingly.

Also, when editing a user info or managing access the URL contains
the host information. If no host information is found a '-' is
used instead.

Change-Id: I4ce235e59ae36de9bb6b87f0a25ac0cb8f537816
Closes-Bug: DBAAS-2370