Horizon: Pass force delete flag to client
After limiting reset status to be only accessible by an admin it
was required by the client to pass a flag to the reset status API
if a force delete action is being performed. The force delete
flag indicates that the reset status action can be performed
even though the user may not be an admin.

Add the force delete flag to the instance and cluster reset status
call.

Change-Id: I4c0fcaa07e137318d2439a115c5c445a57e16e88
Closes-Bug: DBAAS-2387